DCAT-2: PR comments
